Replace all doubles with decimal for future experiments with number types.

+++ src/abstractsyntaxtreebuilder.h
@@ -20,6 +20,7 @@
 #define ABSTRACTSYNTAXTREEBUILDER_H
 
 #include <QList>
+#include "decimal.h"
 #include "declaration.h"
 #include "statement.h"
 #include "variable.h"
@@ -90,7 +91,7 @@ public:
 	virtual Argument* buildArgument(Variable*,Expression*)=0;
 	virtual Expression* buildLiteral()=0;
 	virtual Expression* buildLiteral(bool)=0;
-	virtual Expression* buildLiteral(double value)=0;
+	virtual Expression* buildLiteral(decimal value)=0;
 	virtual Expression* buildLiteral(QString* value)=0;
 	virtual Variable* buildVariable(QString* name)=0;
 	virtual Expression* buildVariable(Variable*)=0;
