+++ src/function/degfunction.cpp
@@ -29,7 +29,7 @@ Value* DegFunction::evaluate(Context* ctx)
 {
 	NumberValue* numVal=dynamic_cast<NumberValue*>(getParameterArgument(ctx,0));
 	if(numVal) {
-		double num=numVal->getNumber();
+		decimal num=numVal->getNumber();
 
 		return new NumberValue(num*360.0/M_TAU);
 	}
