+++ src/module/rotatemodule.h
@@ -28,10 +28,10 @@ public:
 	RotateModule();
 	Node* evaluate(Context*);
 private:
-	bool rightAngle(double);
-	double round(double);
-	double hardSin(double);
-	double hardCos(double);
+	bool rightAngle(decimal);
+	decimal round(decimal);
+	decimal hardSin(decimal);
+	decimal hardCos(decimal);
 };
 
 #endif // ROTATEMODULE_H
